Add files via upload
Here I demonstrate using Postgres. Aggregation, filtering, CTE, subqueries, window functions, merging etc.

The aim of the project is to study the influence of the characteristics of players and their game characters on the purchase of in-game currency "paradise petals", as well as to evaluate the activity of players when making in-game purchases.

Selling "paradise petals" is the main source of income for the development team. The game team plans to attract paying players and promote the purchase of epic items through advertising.

Several tasks needed to be solved:

find out what proportion of players buy the in-game currency "paradise petals" for real money and whether the proportion of paying players depends on the character's race;
study in detail how the purchase of epic items occurs in the game.